fix handleSubmit
